refactor: update project folder structure
