+++ .pre-commit-config.yaml
@@ -8,7 +8,7 @@ repos:
   - hooks:
       - id: fmt
       - id: cargo-check
-      - id: clippy
+      # - id: clippy
     repo: https://github.com/doublify/pre-commit-rust
     rev: v1.0
   - repo: https://github.com/alessandrojcm/commitlint-pre-commit-hook
