+++ src/components/style/mod.rs
@@ -1,0 +1,7 @@
+mod border;
+mod box_style;
+mod text;
+
+pub use border::BorderStyle;
+pub use box_style::{Alignment, BoxStyle, FlexBox, Padding};
+pub use text::TextStyle;
