+++ src/components/widgets/mod.rs
@@ -1,0 +1,3 @@
+mod button;
+
+pub use button::Button;
